Use passed-in fetch for phase 1 of the 402 request
Previously we used the global fetch for phase 1, because there was
no guarantee that the passed-in fetch would properly handle 402s
(e.g. it could error before our handlers could act).  This ended
up being counterintuitive in practice.

This changes the default to the passed in fetch, while still
providing an option to be explicit about phase 1.